change bittiger to bitTiger and adsId type int to long

+++ src/main/java/com/bitTiger/searchAds/main/AdSearchEndpoints.java
@@ -3,7 +3,7 @@ package com.bitTiger.searchAds.main;
 import java.util.List;
 import java.util.Map;
 
-import static com.bittiger.searchAds.service.OfyService.ofy;
+import static com.bitTiger.searchAds.service.OfyService.ofy;
 
 import com.bitTiger.searchAds.adsInfo.AdsInfo;
 import com.bitTiger.searchAds.adsInfo.AdsInventory;
@@ -19,7 +19,7 @@ import com.bitTiger.searchAds.index.AdsIndexImpl;
 import com.bitTiger.searchAds.queryParser.QueryParser;
 import com.bitTiger.searchAds.queryParser.QueryParserImpl;
 
-import com.bittiger.searchAds.datastore.AdsData;
+import com.bitTiger.searchAds.datastore.AdsData;
 
 import com.google.api.server.spi.config.Api;
 import com.google.api.server.spi.config.ApiMethod;
@@ -54,7 +54,7 @@ public class AdSearchEndpoints {
     public List<String> getTokens(@Named("name") String queryString) {
     
     	// add new test data
-    	AdsData data = new AdsData(99, 999);
+    	AdsData data = new AdsData((long) 99, 999);
     	ofy().save().entity(data).now();
     	
         return QUERY_PARSER.parseQuery(queryString);
